# backend/src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: QuantEnergx API
  description: |
    Industry-disrupting energy trading platform with advanced features across 13 major categories.
    
    ## Features
    - Real-time trading infrastructure
    - Multi-commodity support (Oil, Gas, Electricity, Renewables, Carbon Credits)
    - Derivatives trading (Futures, Options, Swaps, Structured Notes)
    - Advanced risk management with VaR calculations
    - Multi-exchange connectivity (ICE, EEX, CME, NYMEX, DME, OPEC)
    - Compliance across multiple jurisdictions (MiFID II, REMIT, Dodd-Frank, EMIR)
    - Settlement processing (Cash/Physical/Net)
    - Plugin architecture and webhook integrations
    - OCR document processing
    - IoT and SCADA integration
    
  version: 2.0.0
  contact:
    name: QuantEnergx Team
    url: https://github.com/akramahmed1/quantenergx
    email: support@quantenergx.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.quantenergx.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - timestamp
...
paths:
...
  /api/v1/plugins/{name}/execute:
    post:
      tags:
        - Plugins
      summary: Execute plugin
      description: Execute a specific plugin with input data
      parameters:
        - name: name
          in: path
          required: true
          description: Plugin name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  description: Plugin-specific input data
              required:
                - input
      responses:
        '200':
          description: Plugin executed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          result:
                            type: object
                          executionTime:
                            type: number
                          timestamp:
                            type: string
                            format: date-time

  # Webhook Endpoints
...
  /api/v1/webhooks/{type}:
    post:
      tags:
        - Webhooks
      summary: Process webhook
      description: Process incoming webhook from external service
      parameters:
        - name: type
          in: path
          required: true
          description: Webhook type
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                source:
                  type: string
                data:
                  type: object
                signature:
                  type: string
              required:
                - id
                - data
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          processed:
                            type: boolean
                          webhookId:
                            type: string

  # WebSocket Statistics
...